+++ plugin_tooling/src-tests/com/googlecode/goclipse/tooling/CommonGoToolingTest.java
@@ -27,7 +27,7 @@ public class CommonGoToolingTest extends CommonTestExt {
 	
 	public static final GoRoot SAMPLE_GO_ROOT = new GoRoot(TESTS_WORKDIR.resolve("goRoot").toString());
 	
-	public static final Path SAMPLE_GOPATH_Entry = TESTS_WORKDIR.resolve("goPath");
+	public static final Path SAMPLE_GOPATH_Entry = TESTS_WORKDIR.resolve("goPathEntry");
 	public static final GoPath SAMPLE_GO_PATH = new GoPath(SAMPLE_GOPATH_Entry.toString());
 	
 	protected static final GoEnvironment SAMPLE_GOEnv_1 = new GoEnvironment(