+++ documentation/ChangeLog.md
@@ -1,9 +1,11 @@
 ## release ChangeLog
 
-### Goclipse 0.8.x
+### Goclipse 0.9.0
+ * Eclipse Go projects can now be created inside the "src" folder of a GOPATH entry, and that scenario will be handled properly by the toolchain (gocode, oracle, build, etc.).
+   * In this case, the project location will not be implicitly added to the GOPATH, as the project is part of the GOPATH already. 
  * Fixed: Project Explorer no longer shows GOPATH entry if that entry location is the as the project location.
  * Added: GOROOT entry in Project Explorer now sorted at the top, and has own icon.
- * Added: GOPATH entries have location subtext. ###
+ * Added: Project Explorer: GOPATH entries have location subtext. ###
  * Added gocode process argument information to gocode log console.
  * Added Ctrl-click open definition functionality using `go oracle`.
    * Path to `oracle` command can be configured in the Go preference page.